third_party/WebKit/LayoutTests/fast/canvas/draw-system-focus-ring-expected.txt

PASS canvas.drawSystemFocusRing(null); did not throw exception.
PASS canvas.drawSystemFocusRing(); threw exception TypeError: Failed to execute 'drawSystemFocusRing' on 'CanvasRenderingContext2D': 1 argument required, but only 0 present..
PASS canvas.drawSystemFocusRing(null); did not throw exception.
PASS data[0] != 0 || data[1] != 0 || data[2] != 0 is true
PASS data[0] == 0 && data[1] == 0 && data[2] == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
